Move source files to src/
