+++ src/bin/robinhood-to-csv.js
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 
-const pkg = require('../package.json')
+const pkg = require('../../package.json')
 const program = require('commander')
 
 const main = require('../lib/main')
