+++ src/lib/main.js
